file restructuring as a group
