+++ api/add_bet_to_db.php
