+++ api/retrieve_bet_history.php
