+++ api/retrieve_game_data.php
@@ -1,10 +1,7 @@
 <?php
 require_once('mysql_connect.php');
 date_default_timezone_set('UTC');
-//$league = 'NFL';
-
 //print(json_encode($_POST));     //for testing purposes
-
 //collect league from index
 $league = $_POST['league'];  //may not need to upper
 //lookup the league in league table
@@ -19,10 +16,6 @@ if(mysqli_num_rows($result)){
 
 //create empty array (of object)
 $data = [];
-//$league_id = $_POST[$league_id];
-//$league_id = 889;
-//$league_id = 487;
-
 //full data attempt
 $game_query = "SELECT g.ID AS game_id, a.full_name AS full_name_a, a.abbr_name AS abbr_name_a, a.logo_src AS logo_src_a, h.full_name AS full_name_h, h.abbr_name AS abbr_name_h, h.logo_src AS logo_src_h, g.home_spread, g.spread_odds_a, g.spread_odds_h, g.moneyline_odds_a, g.moneyline_odds_h, g.overunder_points, g.overunder_odds_o, g.overunder_odds_u, g.game_time FROM `games` AS g JOIN `teams` AS h ON g.team_h_id = h.ID JOIN `teams` AS a ON g.team_a_id = a.ID WHERE g.league_id = $league_id ORDER BY g.game_time";
 $result = mysqli_query($conn, $game_query);
