+++ api/test_calls/cfb_odds.php
@@ -1,5 +1,5 @@
 <?php 
-require('odds_call.php');
+require('../api_calls.php');
 $stuff = odds_call(15,880);
 
 echo "<pre>";
