+++ api/test_calls/mlb_odds.php
@@ -1,5 +1,5 @@
 <?php 
-require('odds_call.php');
+require('../api_calls.php');
 $stuff = odds_call(3,246);
 
 echo "<pre>";
