+++ api/test_calls/nba_odds.php
@@ -1,12 +1,10 @@
-<?php 
+<?php
 $connection = mysqli_connect('localhost', 'root', '', 'betting');
 
-require('api_calls.php');
-$stuff = odds_call(15,889);
+require('../api_calls.php');
+$stuff = odds_call(4,487);
 
 echo "<pre>";
 print_r($stuff);
 echo "</pre>";
-
-
  ?>
