+++ api/test_calls/nfl_odds.php
@@ -1,10 +1,12 @@
-<?php
+<?php 
 $connection = mysqli_connect('localhost', 'root', '', 'betting');
 
-require('api_calls.php');
-$stuff = odds_call(4,487);
+require('../api_calls.php');
+$stuff = odds_call(15,889);
 
 echo "<pre>";
 print_r($stuff);
 echo "</pre>";
+
+
  ?>
