+++ api/test_calls/nhl_odds.php
@@ -1,5 +1,5 @@
 <?php
-require('api_calls.php');
+require('../api_calls.php');
 $stuff = odds_call(19,1460);
 
 echo "<pre>";
