--- bet_data.html
@@ -1,5 +1,0 @@
-<div ng-repeat=''>
-    <div class = 'game'>
-        {{date_data}}
-    </div>
-</div>
