--- game_to_bet.html
@@ -1,9 +1,0 @@
-
-<div ng-repeat='bet_data in sc.arr'>
-    <div class = 'game'>
-        {{bet_data.game_data}}
-    </div>
-    <div class = 'bet'>
-        {{bet_data.bet_data}}
-    </div>
-</div>
